refactor: contact作成処理をリソースルートとして切り出し
