+++ app/routes/__.poc.contacts/route.tsx
@@ -1,4 +1,4 @@
-import { Link, Outlet, redirect, useNavigation } from 'react-router';
+import { Link, Outlet, useNavigation } from 'react-router';
 import { prisma } from '~/.server/lib/prisma-client';
 import {
   Breadcrumb,
@@ -34,28 +34,6 @@ export const loader = async ({ request }: Route.LoaderArgs) => {
       : true,
   );
   return { contacts: filteredContacts, q };
-};
-
-export const action = async ({ request }: Route.ActionArgs) => {
-  const formData = await request.formData();
-  const { _action } = Object.fromEntries(formData);
-
-  switch (_action) {
-    case 'new': {
-      const newContact = {
-        first: '',
-        last: '',
-        avatar: '',
-        twitter: '',
-        notes: '',
-        favorite: false,
-      };
-
-      const res = await prisma.contact.create({ data: newContact });
-      // 編集ページにリダイレクト
-      return redirect(`/poc/contacts/${res.id}/edit`);
-    }
-  }
 };
 
 const ContactsLayout = ({ loaderData }: Route.ComponentProps) => {
